build-data/requirements.txt
pandas
requests
openpyxl
psycopg2-binary
python-dateutil
